+++ ci4/app/Interfaces/PostCommand.php
@@ -3,6 +3,7 @@
  * @property string $name
  * @property string $command
  * @property bool $allPods
+ * @property string $container
  */
 interface PostCommand {