+++ vue/src/components/Dialogs/Setup/DeploymentSpecifications/UpdateDialogs/DeploymentSpecificationUpdatePostCommandDialog.vue
@@ -7,6 +7,7 @@ export interface DeploymentSpecificationUpdatePostCommandDialog_Input {
         name: string;
         command: string;
         allPods: boolean;
+        container: string;
     };
 
     onSaveCallback: () => void;
@@ -19,6 +20,7 @@ const showDialog = ref(false);
 const name = ref('');
 const command = ref('');
 const allPods = ref(false);
+const container = ref('');
 
 // <editor-fold desc="Functions">
 
@@ -37,6 +39,7 @@ function render() {
     name.value = props.input.postCommand.name ?? '';
     command.value = props.input.postCommand.command ?? '';
     allPods.value = props.input.postCommand.allPods ?? false;
+    container.value = props.input.postCommand.container ?? '';
     showDialog.value = true;
 }
 
@@ -53,6 +56,7 @@ function onSaveBtnClicked() {
     props.input.postCommand.name = name.value;
     props.input.postCommand.command = command.value;
     props.input.postCommand.allPods = allPods.value;
+    props.input.postCommand.container = container.value;
     props.input.onSaveCallback();
     close();
 }
@@ -89,6 +93,14 @@ function onCloseBtnClicked() {
                             v-model="command"
                             variant="outlined"
                             label="Command"
+                            hide-details
+                        />
+                    </v-col>
+                    <v-col cols="12">
+                        <v-text-field
+                            v-model="container"
+                            variant="outlined"
+                            label="Container"
                             hide-details
                         />
                     </v-col>